study_summary/requirements.txt
# Study Summary Automation
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
notebooklm-py>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
slack-sdk>=3.23.0
